Information Content Steps

This function estimates the information content of a character char when e extra steps are present, for all possible values of e.

ICSteps(char, ambiguousToken = 0, expectedMinima = 25,
  maxIter = 10000, warn = TRUE)

Arguments

char
The character in question.

ambiguousToken
Which token, if any, corresponds to the ambiguous token (?) (not yet fully implemented).

expectedMinima
sample enough trees that the rarest step counts is expected to be seen at least this many times.

maxIter
Maximum iterations to conduct.

warn
Boolean (TRUE/FALSE): display warnings when concavity functions are generated by approximation.

Details

Calculates the number of trees consistent with the character having e extra steps, where e ranges from its minimum possible value (i.e. number of different tokens minus one) to its maximum. The number of trees with no extra steps can be calculated exactly; the number of trees with more additional steps must be approximated. The function samples n.iter trees, or enough trees that the trees with the minimum number of steps will be recovered at least expected.minima times, in order to obtain precise results.

References

Faith DP, Trueman JWH (2001). “Towards an inclusive philosophy for phylogenetic inference.” Systematic Biology, 50(3), 331--350. doi: 10.1080/10635150118627 .

Examples

{
  # A character that is present in ten taxa and absent in five
  character <- c(rep(1, 10), rep(2, 5))
  ICSteps (character)
}

#> Warning: Will truncate number of iterations at maxIter = 10000

#>   Token count510=11.09bits @ 0 extra steps; simulating10000trees to estimate cost of further steps.

#>            1            2            3            4            5 
#> 0.0004576659 0.0112970878 0.0960752334 0.3746034333 0.5175665797