Nearest Neighbour Interchange (NNI)

Performs a single iteration of the nearest-neighbour interchange algorithm. Based on the corresponding phangorn function, but re-coded to improve speed.

NNI(tree, edgeToBreak = NULL)

NNISwap(parent, child, nTips = (length(parent)/2L) + 1L, edgeToBreak = NULL)

RootedNNI(tree, edgeToBreak = NULL)

RootedNNISwap(
  parent,
  child,
  nTips = (length(parent)/2L) + 1L,
  edgeToBreak = NULL
)

Arguments

tree
A tree of class phylo.

edgeToBreak
(optional) integer specifying the index of an edge to bisect/prune, generated randomly if not specified. Alternatively, set to -1 to return a complete list of all trees one step from the input tree.

parent
Integer vector corresponding to the first column of the edge matrix of a tree of class phylo, i.e. tree$edge[, 1].

child
Integer vector corresponding to the second column of the edge matrix of a tree of class phylo, i.e. tree$edge[, 2].

nTips
(optional) Number of tips.

Value

Returns a tree with class phylo (if returnAll = FALSE) or a set of trees, with class multiPhylo (if returnAll = TRUE).

a list containing two elements, corresponding in turn to the rearranged parent and child parameters

a list containing two elements, corresponding in turn to the rearranged parent and child parameters

Details

Branch lengths are not supported.

All nodes in a tree must be bifurcating; ape::collapse.singles and ape::multi2di may help.

Functions

NNISwap: faster version that takes and returns parent and child parameters

RootedNNI: Perform NNI rearrangement, retaining position of root

RootedNNISwap: faster version that takes and returns parent and child parameters

References

The algorithm is summarized in Felsenstein J (2004). Inferring phylogenies. Sinauer Associates, Sunderland, Massachusetts.

See also

Examples

tree <- ape::rtree(20, br = NULL)
NNI(tree)

#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.

NNI(tree, edgeToBreak = -1)

#> [[1]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[2]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[3]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[4]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[5]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[6]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[7]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[8]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[9]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[10]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[11]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[12]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[13]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[14]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[15]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[16]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[17]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[18]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[19]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[20]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[21]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[22]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[23]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[24]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[25]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[26]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[27]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[28]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[29]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[30]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[31]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[32]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[33]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[34]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[35]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#> 
#> [[36]]
#> 
#> Phylogenetic tree with 20 tips and 19 internal nodes.
#> 
#> Tip labels:
#> 	t13, t20, t3, t19, t2, t7, ...
#> 
#> Rooted; no branch lengths.
#>