Rearrange edges of a phylogenetic tree

RearrangeEdges() performs the specified edge rearrangement on a matrix that corresponds to the edges of a phylogenetic tree, returning the score of the new tree. Will generally be called from within a tree search function.

RearrangeEdges(
  parent,
  child,
  dataset,
  TreeScorer = MorphyLength,
  EdgeSwapper,
  scoreToBeat = TreeScorer(parent, child, dataset, ...),
  iter = "?",
  hits = 0L,
  verbosity = 0L,
  ...
)

Arguments

parent
Integer vector corresponding to the first column of the edge matrix of a tree of class phylo, i.e. tree$edge[, 1].

child
Integer vector corresponding to the second column of the edge matrix of a tree of class phylo, i.e. tree$edge[, 2].

dataset
Third argument to pass to TreeScorer.

TreeScorer
function to score a given tree. The function will be passed three parameters, corresponding to the parent and child entries of a tree's edge list, and a dataset.

EdgeSwapper
a function that rearranges a parent and child vector, and returns a list with modified vectors; for example SPRSwap().

scoreToBeat
Double giving score of input tree.

iter
iteration number of calling function, for reporting to user only.

hits
Integer giving number of times the input tree has already been hit.

verbosity
Level of detail to display in console: larger numbers provide more verbose feedback to the user.

...
further arguments to pass to TreeScorer() (e.g. TipsAreColumns, dataset).

Value

This function returns a list with two to four elements, corresponding to a binary tree: - 1. Integer vector listing the parent node of each edge; - 2. Integer vector listing the child node of each edge; - 3. Score of the tree; - 4. Number of times that score has been hit.

Details

RearrangeTree() performs one tree rearrangement of a specified type, and returns the score of the tree (with the given dataset). It also reports the number of times that this score was hit in the current function call.

Examples

data('Lobo', package='TreeTools')
tree <- TreeTools::NJTree(Lobo.phy)
edge <- tree$edge
parent <- edge[, 1]
child <- edge[, 2]
dataset <- PhyDat2Morphy(Lobo.phy)
RearrangeEdges(parent, child, dataset, EdgeSwapper = RootedNNISwap)

#> [[1]]
#>  [1] 49 50 51 52 53 54 55 56 56 57 57 55 54 53 58 59 60 60 59 61 61 58 52 62 63
#> [26] 64 65 65 64 66 66 63 67 68 69 70 71 72 73 74 74 73 72 71 70 69 75 76 77 77
#> [51] 76 78 79 79 78 80 80 75 68 81 81 67 62 82 83 84 85 86 86 85 84 87 88 89 90
#> [76] 90 89 88 87 83 91 92 92 91 93 94 94 93 95 95 82 51 50 49
#> 
#> [[2]]
#>  [1] 50 51 52 53 54 55 56  1 57  2  3  4  5 58 59 60  6  7 61  8  9 10 62 63 64
#> [26] 65 11 12 66 13 14 67 68 69 70 71 72 73 74 15 16 17 18 19 20 75 76 77 21 22
#> [51] 78 79 23 24 80 25 26 27 81 28 29 30 82 83 84 85 86 31 32 33 87 88 89 90 34
#> [76] 35 36 37 38 91 92 39 40 93 94 41 42 95 43 44 45 46 47 48
#> 
#> [[3]]
#> [1] 349
#> 
#> [[4]]
#> [1] 0
#>