Morphy object from single character

Morphy object from single character

Usage

SingleCharMorphy(char, gap = "inapp")

Arguments

char

State of each character at each tip in turn, in a format that will be converted to a character string by paste0(char, ";", collapse="").

gap

An unambiguous abbreviation of inapplicable, ambiguous (= missing), or extra state, specifying how gaps will be handled.

Value

A pointer to an object of class morphyObj. Don't forget to unload it when you've finished with it.

See also

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

morphyObj <- SingleCharMorphy("-0-0", gap = "Extra")
RandomTreeScore(morphyObj)
#> [1] 2
morphyObj <- UnloadMorphy(morphyObj)