Tree Search using Successive Approximations

Searches for a tree that is optimal under the Successive Approximations criterion.

SuccessiveApproximations(
  tree,
  dataset,
  outgroup = NULL,
  k = 3,
  maxSuccIter = 20,
  ratchetHits = 100,
  searchHits = 50,
  searchIter = 500,
  ratchetIter = 5000,
  verbosity = 0,
  suboptimal = 0.1
)

Arguments

tree
A tree of class phylo.

dataset
A phylogenetic data matrix of class phyDat, whose names correspond to the labels of any accompanying tree.

outgroup
if not NULL, taxa on which the tree should be rooted

k
Constant for successive approximations, see Farris 1969 p. 379

maxSuccIter
maximum iterations of successive approximation

ratchetHits
maximum hits for parsimony ratchet

searchHits
maximum hits in tree search

searchIter
maximum iterations in tree search

ratchetIter
maximum iterations of parsimony ratchet

verbosity
integer (default 0) specifying how much detail to print to stdout

suboptimal
retain trees that are this proportion less optimal than the optimal tree

Value

SuccessiveApproximations() returns a list of class multiPhylo containing optimal (and slightly suboptimal, if suboptimal > 0) trees.