Retrieves the current list of symbols.

Returns a pointer to the string of character state symbols currently being used by Morphy (i.e. either the list of symbols extracted from the matrix, or the caller-specified values).

mpl_get_symbols(morphyobj)

Arguments

morphyobj

An instance of the Morphy object.

Value

A C-style (null-terminated) string of the character state symbols being used. NULL if failure.

See also

Author

Martin Brazeau