Updates the nodal sets for a lower ("dummy") root node

If trees are rooted, then Morphy uppass functions require a lower or "dummy" root in order to function properly. This function should be called to set the nodal state sets to the dummy root. The nodal set will be equal to the set of the root node, unless there is an ambiguous union of applicable and gap tokens when gaps are treated as in applicable. In which case, the set union is resolved in favour of any applicable tokens in the set.

Usage

mpl_update_lower_root(l_root_id, root_id, morphyobj)

Arguments

l_root_id

The index of the lower root.

root_id

The index of the upper root node.

morphyobj

An instance of the Morphy object.

Value

A Morphy error code.

See also

Author

Thomas Guillerme