Find most parsimonious trees

Search for most parsimonious trees using the parsimony ratchet and TBR rearrangements, treating inapplicable data as such using the algorithm of Brazeau, Guillerme & Smith (2019).

Tree search will be conducted from a specified or automatically-generated starting tree in order to find a tree with an optimal parsimony score, under implied or equal weights, treating inapplicable characters as such in order to avoid the artefacts of the standard Fitch algorithm (see Maddison 1993; Brazeau et al. 2019).

MaximizeParsimony(
  dataset,
  tree = NJTree(dataset),
  ratchIter = 12L,
  tbrIter = 6L,
  finalIter = 3L,
  maxHits = 20L,
  concavity = Inf,
  constraint = NULL,
  verbosity = 2L,
  session = NULL
)

EasyTrees()

EasyTreesy()

Arguments

dataset
A phylogenetic data matrix of class phyDat, whose names correspond to the labels of any accompanying tree.

tree
(optional) A bifurcating tree of class phylo, containing only the tips listed in dataset, from which the search should begin. If unspecified, a neighbour-joining tree will be generated from dataset. Edge lengths are not supported and will be deleted.

ratchIter
Numeric specifying number of iterations of the parsimony ratchet (Nixon 1999) to conduct.

tbrIter
Numeric specifying the maximum number of TBR break points to evaluate before concluding each search. The counter is reset to zero each time tree score improves. One 'iteration' comprises breaking a single branch and evaluating all possible reconnections.

finalIter
Numeric: the final round of tree search will evaluate finalIter × tbrIter TBR break points.

maxHits
Numeric specifying the maximum times that an optimal parsimony score may be hit before concluding a ratchet iteration or final search concluded.

concavity
Numeric specifying concavity constant for implied step weighting; set as Inf for equal step weights (which is a bad idea; see Smith 2019).

constraint
Either NULL or an object of class phyDat. Trees that are not perfectly compatible with each character in constraint will not be considered during search. See vignette for further examples.

verbosity
Integer specifying level of messaging; higher values give more detailed commentary on search progress. Set to 0 to run silently.

session
'shiny' session identifier to allow setProgress() calls to be sent when MaximizeParsimony() is called from within a shiny app..

Details

Tree search commences with ratchIter iterations of the parsimony ratchet (Nixon 1999), which bootstraps the input dataset in order to escape local optima. A final round of tree bisection and reconnection (TBR) is conducted to broaden the sampling of trees.

This function can be called using the R command line / terminal, or through the 'shiny' graphical user interface app (type EasyTrees() to launch).

For detailed documentation of the TreeSearch package, including full instructions for loading phylogenetic data into R and initiating and configuring tree search, see the package documentation.

References

Brazeau MD, Guillerme T, Smith MR (2019). “An algorithm for morphological phylogenetic analysis with inapplicable data.” Systematic Biology, 68(4), 619--631. doi: 10.1093/sysbio/syy083 , https://doi.org/10.1093/sysbio/syy083.

Maddison WP (1993). “Missing data versus missing characters in phylogenetic analysis.” Systematic Biology, 42(4), 576--581. doi: 10.1093/sysbio/42.4.576 , https://doi.org/10.1093/sysbio/42.4.576.

Nixon KC (1999). “The Parsimony Ratchet, a new method for rapid parsimony analysis.” Cladistics, 15(4), 407--414. ISSN 0748-3007, doi: 10.1111/j.1096-0031.1999.tb00277.x , https://doi.org/10.1111/j.1096-0031.1999.tb00277.x.

Smith MR (2019). “Bayesian and parsimony approaches reconstruct informative trees from simulated morphological datasets.” Biology Letters, 15(2), 20180632. doi: 10.1098/rsbl.2018.0632 , https://doi.org/10.1098/rsbl.2018.0632.

Author

Martin R. Smith (martin.smith@durham.ac.uk)

Examples

library('TreeTools')
data('Lobo', package = 'TreeTools')
dataset <- Lobo.phy

# A very quick run for demonstration purposes
MaximizeParsimony(dataset, ratchIter = 0, tbrIter = 1, concavity = 10,
                  maxHits = 5, verbosity = 4)

#> - Parsimony search with 0 ratchet iterations; TBR depth 1; 5 hits; k = 10.
#>   2021-03-03 19:58:38
#>   Initial score: 9.1938

#> - Final TBR search.

#>    - New TBR iteration (depth 1)

#>   - 1 TBR rearrangements found score 8.78255 5 times.

#> - Final score: 8.78255078255079
#> 

#> 5 phylogenetic trees

# Be sure to check that the score has converged on a global optimum,
# conducting additional iterations and runs as necessary.

if (FALSE)  # launches 'shiny' point-and-click interface
  EasyTrees() 

# Tree search with a constraint
constraint <- MatrixToPhyDat(c(a = 1, b = 1, c = 0, d = 0, e = 0, f = 0))
characters <- MatrixToPhyDat(matrix(
  c(0, 1, 1, 1, 0, 0,
    1, 1, 1, 0, 0, 0), ncol = 2,
  dimnames = list(letters[1:6], NULL)))
MaximizeParsimony(characters, constraint = constraint)

#>  - Looking for a tree that is consistent with `constraint`...

#> - Parsimony search with 12 ratchet iterations; TBR depth 6; 20 hits; k = Inf.
#>   2021-03-03 19:58:39
#>   Initial score: 3

#> - Performing parsimony ratchet.

#>  - Ratchet iteration 1: Search from bootstrapped dataset.

#>  - Ratchet iteration 1: Search using original data

#>  - Ratchet iteration 1 hit best score 3 again

#>  - Ratchet iteration 2: Search from bootstrapped dataset.

#>  - Ratchet iteration 2: Search using original data

#>  - Ratchet iteration 2 hit best score 3 again

#>  - Ratchet iteration 3: Search from bootstrapped dataset.

#>  - Ratchet iteration 3: Search using original data

#>  - Ratchet iteration 3 hit best score 3 again

#>  - Ratchet iteration 4: Search from bootstrapped dataset.

#>  - Ratchet iteration 4: Search using original data

#>  - Ratchet iteration 4 hit best score 3 again

#>  - Ratchet iteration 5: Search from bootstrapped dataset.

#>  - Ratchet iteration 5: Search using original data

#>  - Ratchet iteration 5 did not hit best score 3.

#>  - Ratchet iteration 6: Search from bootstrapped dataset.

#>  - Ratchet iteration 6: Search using original data

#>  - Ratchet iteration 6 did not hit best score 3.

#>  - Ratchet iteration 7: Search from bootstrapped dataset.

#>  - Ratchet iteration 7: Search using original data

#>  - Ratchet iteration 7 hit best score 3 again

#>  - Ratchet iteration 8: Search from bootstrapped dataset.

#>  - Ratchet iteration 8: Search using original data

#>  - Ratchet iteration 8 hit best score 3 again

#>  - Ratchet iteration 9: Search from bootstrapped dataset.

#>  - Ratchet iteration 9: Search using original data

#>  - Ratchet iteration 9 hit best score 3 again

#>  - Ratchet iteration 10: Search from bootstrapped dataset.

#>  - Ratchet iteration 10: Search using original data

#>  - Ratchet iteration 10 hit best score 3 again

#>  - Ratchet iteration 11: Search from bootstrapped dataset.

#>  - Ratchet iteration 11: Search using original data

#>  - Ratchet iteration 11 hit best score 3 again

#>  - Ratchet iteration 12: Search from bootstrapped dataset.

#>  - Ratchet iteration 12: Search using original data

#>  - Ratchet iteration 12 hit best score 3 again

#> - Final TBR search.

#> - Final score: 3
#> 

#> 1 phylogenetic tree